--- pkgs/edge-worker/src/flow/createFlowWorkerV2.ts
@@ -1,175 +1,0 @@
-import type { AnyFlow } from '@pgflow/dsl';
-import { ExecutionController } from '../core/ExecutionController.js';
-import { StepTaskPollerV2, type StepTaskPollerV2Config } from './StepTaskPollerV2.js';
-import { StepTaskExecutor } from './StepTaskExecutor.js';
-import { PgflowSqlClient } from '@pgflow/core';
-import { Queries } from '../core/Queries.js';
-import type { StepTaskRecord } from './types.js';
-import type { IExecutor } from '../core/types.js';
-import type { Logger } from '../platform/types.js';
-import { Worker } from '../core/Worker.js';
-import postgres from 'postgres';
-import { FlowWorkerLifecycle } from './FlowWorkerLifecycle.js';
-import { BatchProcessor } from '../core/BatchProcessor.js';
-import { randomUUID } from 'crypto';
-
-/**
- * Configuration for the flow worker V2 with two-phase polling
- */
-export type FlowWorkerV2Config = {
-  /**
-   * How many tasks are processed at the same time
-   * @default 10
-   */
-  maxConcurrent?: number;
-
-  /**
-   * PostgreSQL connection string.
-   * If not provided, it will be read from the EDGE_WORKER_DB_URL environment variable.
-   */
-  connectionString?: string;
-
-  /**
-   * Optional SQL client instance
-   */
-  sql?: postgres.Sql;
-
-  /**
-   * How many connections to the database are opened
-   * @default 4
-   */
-  maxPgConnections?: number;
-
-  /**
-   * Batch size for polling messages
-   * @default 10
-   */
-  batchSize?: number;
-
-  /**
-   * Visibility timeout for messages in seconds
-   * @default 2
-   */
-  visibilityTimeout?: number;
-
-  /**
-   * In-worker polling interval in seconds
-   * @default 2
-   */
-  maxPollSeconds?: number;
-
-  /**
-   * In-database polling interval in milliseconds
-   * @default 100
-   */
-  pollIntervalMs?: number;
-
-  /**
-   * Worker ID for tracking task ownership
-   * If not provided, a random UUID will be generated
-   */
-  workerId?: string;
-};
-
-/**
- * Creates a new Worker instance for processing flow tasks using the two-phase polling approach.
- * This eliminates race conditions by separating message polling from task processing.
- *
- * @param flow - The Flow DSL definition
- * @param config - Configuration options for the worker
- * @returns A configured Worker instance ready to be started
- */
-export function createFlowWorkerV2<TFlow extends AnyFlow>(
-  flow: TFlow,
-  config: FlowWorkerV2Config,
-  createLogger: (module: string) => Logger
-): Worker {
-  const logger = createLogger('createFlowWorkerV2');
-
-  // Create abort controller for graceful shutdown
-  const abortController = new AbortController();
-  const abortSignal = abortController.signal;
-
-  if (!config.sql && !config.connectionString) {
-    throw new Error(
-      "Either 'sql' or 'connectionString' must be provided in FlowWorkerV2Config."
-    );
-  }
-
-  const sql =
-    config.sql ||
-    postgres(config.connectionString!, {
-      max: config.maxPgConnections,
-      prepare: false,
-    });
-
-  // Create the pgflow adapter
-  const pgflowAdapter = new PgflowSqlClient<TFlow>(sql);
-
-  // Use flow slug as queue name, or fallback to 'tasks'
-  const queueName = flow.slug || 'tasks';
-  logger.debug(`Using queue name: ${queueName}`);
-
-  // Generate worker ID if not provided
-  const workerId = config.workerId || randomUUID();
-  logger.debug(`Using worker ID: ${workerId}`);
-
-  // Create specialized FlowWorkerLifecycle with the proxied queue and flow
-  const queries = new Queries(sql);
-  const lifecycle = new FlowWorkerLifecycle<TFlow>(
-    queries,
-    flow,
-    createLogger('FlowWorkerLifecycle')
-  );
-
-  // Create StepTaskPollerV2 with two-phase approach
-  const pollerConfig: StepTaskPollerV2Config = {
-    batchSize: config.batchSize || 10,
-    queueName: flow.slug,
-    workerId: workerId,
-    visibilityTimeout: config.visibilityTimeout || 2,
-    maxPollSeconds: config.maxPollSeconds || 2,
-    pollIntervalMs: config.pollIntervalMs || 100,
-  };
-  const poller = new StepTaskPollerV2<TFlow>(
-    pgflowAdapter,
-    abortSignal,
-    pollerConfig,
-    createLogger('StepTaskPollerV2')
-  );
-
-  // Create executor factory with proper typing
-  const executorFactory = (
-    record: StepTaskRecord<TFlow>,
-    signal: AbortSignal
-  ): IExecutor => {
-    return new StepTaskExecutor<TFlow>(
-      flow,
-      record,
-      pgflowAdapter,
-      signal,
-      createLogger('StepTaskExecutor')
-    );
-  };
-
-  // Create ExecutionController
-  const executionController = new ExecutionController<StepTaskRecord<TFlow>>(
-    executorFactory,
-    abortSignal,
-    {
-      maxConcurrent: config.maxConcurrent || 10,
-    },
-    createLogger('ExecutionController')
-  );
-
-  // Create BatchProcessor
-  const batchProcessor = new BatchProcessor<StepTaskRecord<TFlow>>(
-    executionController,
-    poller,
-    abortSignal,
-    createLogger('BatchProcessor')
-  );
-
-  // Return Worker
-  return new Worker(batchProcessor, lifecycle, sql, createLogger('Worker'));
-}
